moved two example classes to another package
