+++ docs/website/static/javadocs/edu/iu/dsc/tws/examples/internal/bootstrap/ZKJobMasterFinderExample.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.iu.dsc.tws.examples.internal.bootstrap.ZKJobMasterFinderExample</li>
+<li>edu.iu.dsc.tws.examples.internal.jobmaster.ZKJobMasterFinderExample</li>
 </ul>
 </li>
 </ul>
