+++ docs/website/static/javadocs/edu/iu/dsc/tws/examples/internal/bootstrap/ZKJobMasterRegistrarExample.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.iu.dsc.tws.examples.internal.bootstrap.ZKJobMasterRegistrarExample</li>
+<li>edu.iu.dsc.tws.examples.internal.jobmaster.ZKJobMasterRegistrarExample</li>
 </ul>
 </li>
 </ul>
